% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{choose_genes_tt}
\alias{choose_genes_tt}
\title{Rhythmicity analysis of training data}
\usage{
choose_genes_tt(
  object,
  group1,
  group2,
  group3,
  replicate,
  method = "population",
  parallel = TRUE,
  cores = 6,
  rank_weights = c(pval = 1, rsquared = 1, mrl = 1, amp_cv = 1)
)
}
\arguments{
\item{object}{list containing TimeTeller training model following \code{train_model}}

\item{group1}{if only a subset of Group_1 (Metadata provided) should be used}

\item{group2}{if only a subset of Group_2 (Metadata provided) should be used}

\item{group3}{if only a subset of Group_3 (Metadata provided) should be used}

\item{replicate}{if only a subset of Replicate (Metadata provided) should be used}

\item{method}{method used for rhythmicity analysis. Default is \code{'population'} as in \url{https://tbiomed.biomedcentral.com/articles/10.1186/1742-4682-11-16}}

\item{parallel}{if TRUE, parallel computation using \code{foreach} and \code{doParallel} packages will be used. Default is TRUE}

\item{cores}{if using parallel computation, how many cores should be used. Default is 6}

\item{rank_weights}{named numeric vector of weights for composite ranking. Names must be \code{pval}, \code{rsquared}, \code{mrl}, \code{amp_cv}. Default is equal weighting.}
}
\value{
Returns the updated object with \code{Rhythmicity_Results} data frame
}
\description{
Calculates population cosinor rhythmicity results for the training data,
including phase consistency (MRL) and amplitude reliability (CV of rAMP).
Genes are ranked by a weighted composite of p-value, R-squared, MRL,
and amplitude CV.
}
\references{
Cornelissen, G., 2014. Cosinor-based rhythmometry. Theoretical Biology and Medical Modelling, 11(1), pp.1-24.
}
\author{
Vadim Vasilyev
}
