% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{choose_logthresh}
\alias{choose_logthresh}
\title{Log Threshold selection}
\usage{
choose_logthresh(
  object,
  max_log = 0,
  min_log = -12,
  by_step = -1,
  train_or_test = "test"
)
}
\arguments{
\item{object}{list containing the output of \code{train_model} and \code{test_model} functions}

\item{max_log}{max value of logthresh to be considered. Generally recommended to check \code{[0,-12]} range}

\item{min_log}{min value of logthresh to be considered}

\item{by_step}{step size for logthresh optimisation. We have found -0.5 to be sufficient, however users may benefit from more granular data depending on the application}

\item{train_or_test}{whether the simulation should be run on \code{'train'} or \code{'test'} data}
}
\value{
returns \code{data.frame} containing the results of log threshold selection
}
\description{
This will iterate though multiple values of logthresh and save the relevant metrics for each. Can be used to aid in selecting an appropriate threshold. See \code{?choose_logthresh_plot} for further visualisation
}
\author{
Vadim Vasilyev
}
