% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{choose_logthresh_plot}
\alias{choose_logthresh_plot}
\title{Log Threshold selection plot}
\usage{
choose_logthresh_plot(choose_logthresh_df, cap = 10, perc_flat = 0.02)
}
\arguments{
\item{choose_logthresh_df}{\code{data.frame} containing the output of \code{choose_logthresh} function}

\item{cap}{parameter used for clarity of visualisation. Ratio of LogLik for 1st and 2nd peaks}

\item{perc_flat}{option to target a particular proportion of samples with flat likelihoods (ie their peak is below the currently selected LogThresh)}
}
\value{
returns \code{ggplot} object
}
\description{
Plots the results of \code{choose_logthresh} function
}
\author{
Vadim Vasilyev
}
