% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{display_rhythmicity_results}
\alias{display_rhythmicity_results}
\title{Display the results of cosinor rhythmicity analysis}
\usage{
display_rhythmicity_results(
  object,
  probes_of_interest,
  info_used = "ranks",
  organism
)
}
\arguments{
\item{object}{list containing the output of \code{train_model} and \code{choose_genes_tt} functions}

\item{probes_of_interest}{genes / features of interest to highlight}

\item{info_used}{plot style to output. Must be one of the \code{'ranks'} or \code{'original'}. Choosing \code{'ranks'} allows for more interactivity using \code{ggplotly}}

\item{organism}{this is added for convenience. Will attempt to use \code{gprofiler2::gconvert} to get Entrez gene symbols for more user-friendly visualisation. Therefore, the argument format has to be consistent with that.}
}
\value{
Returns \code{plot_ly} or \code{ggplot} object depending on the \code{info_used} selection
}
\description{
Visualisation of results following \code{choose_genes_tt} function
}
\author{
Vadim Vasilyev
}
