% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Training_functions.R
\name{get_final_likelis}
\alias{get_final_likelis}
\title{Apply log threshold and compute averaged likelihoods}
\usage{
get_final_likelis(object, log_thresh, mode = "train")
}
\arguments{
\item{object}{TimeTeller list object}

\item{log_thresh}{log threshold value}

\item{mode}{either 'train' or 'test'}
}
\value{
Updated object with thresholded and averaged likelihoods
}
\description{
Apply log threshold and compute averaged likelihoods
}
\keyword{internal}
