% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{nature_palette}
\alias{nature_palette}
\title{Nature-style colour palette}
\usage{
nature_palette(n = NULL, type = "qualitative")
}
\arguments{
\item{n}{number of colours to generate (NULL returns the base palette)}

\item{type}{one of 'qualitative', 'sequential_blue', or 'diverging'}
}
\value{
character vector of hex colour codes
}
\description{
Provides qualitative, sequential, or diverging colour palettes suitable
for publication figures.
}
