% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_3d_projection}
\alias{plot_3d_projection}
\title{Visualise the training model}
\usage{
plot_3d_projection(
  object,
  selected_local_projection,
  density = FALSE,
  opacity = 0.05,
  sig_level = 0.9
)
}
\arguments{
\item{object}{list containing TimeTeller training model following \code{train_model}}

\item{selected_local_projection}{training time selected for the local projection}

\item{density}{should covariance matrix cloud be displayed. Default is FALSE}

\item{opacity}{opacity level if \code{density = TRUE}}

\item{sig_level}{confidence level when \code{density = TRUE} and is used to control the size of the ellipsoid. This uses \code{rgl::ellipse3d}}
}
\value{
Returned is the \code{plot_ly} object
}
\description{
Training data projected in the principal component space.
}
\author{
Vadim Vasilyev
}
