% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_cv_res}
\alias{plot_cv_res}
\title{Cross-validation results on the training data}
\usage{
plot_cv_res(cv_object)
}
\arguments{
\item{cv_object}{list containing the output of \code{train_cv} function}
}
\value{
Returns prediction error and theta estimate for the CV groups
}
\description{
Visualisation of results following \code{train_cv} function. Plots prediction error by groups
}
\author{
Vadim Vasilyev
}
