% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_deviation_cv_corrected}
\alias{plot_deviation_cv_corrected}
\title{Cross-validation results on the training data}
\usage{
plot_deviation_cv_corrected(list_cv)
}
\arguments{
\item{list_cv}{list containing the output of \code{train_cv} function}
}
\value{
Returns \code{TimeTeller} prediction error plotted against the actual sample time
}
\description{
Visualisation of results following \code{train_cv} function. Plots prediction error against actual time.
This function can be useful when systematic deviations (eg chronotype) are suspected, in which case error will be adjusted by the median.
Care should be taken otherwise, as this could also mask the algorithm's systematic error, therefore \code{plot_deviation_cv_original} and \code{plot_cv_res} are
the recommended first steps
}
\author{
Vadim Vasilyev
}
