% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_deviation_cv_original}
\alias{plot_deviation_cv_original}
\title{Cross-validation results on the training data}
\usage{
plot_deviation_cv_original(list_cv)
}
\arguments{
\item{list_cv}{list containing the output of \code{train_cv} function}
}
\value{
Returns \code{TimeTeller} prediction error plotted against the actual sample time
}
\description{
Visualisation of results following \code{train_cv} function. Plots prediction error against actual time.
}
\author{
Vadim Vasilyev
}
