% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_genes}
\alias{plot_genes}
\title{Visualise multiple gene expression}
\usage{
plot_genes(object, genes, group1, group2)
}
\arguments{
\item{object}{list containing circadianTT training model following \code{train_model}}

\item{genes}{genes of interest for visualisation}

\item{group1}{groups of interest for plotting. If not supplied, all levels from \code{Group_1} will be plotted}

\item{group2}{groups of interest for plotting. If not supplied, all levels from \code{Group_2} will be plotted}
}
\value{
Returned is the object of class \code{ggplot}
}
\description{
Quick exploratory plot of temporal gene expression across groups. Recommended use when there are no replicates for clarity of visualisation
}
\author{
Vadim Vasilyev
}
