% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_ind_curve}
\alias{plot_ind_curve}
\title{Theta calculation of individual samples}
\usage{
plot_ind_curve(object, sample_num, logthresh, train_or_test = "test")
}
\arguments{
\item{object}{list containing TimeTeller training and test models following \code{train_model} and \code{test_model} respectively}

\item{sample_num}{number of sample to be displayed}

\item{logthresh}{which log threshold should be used for visualisation}

\item{train_or_test}{is the sample coming from \code{train} or \code{test} data}
}
\value{
Returns plots of theta calculation curves produced using standard \code{graphics::plot}
}
\description{
This displays \code{theta} and flat likelihood percentage calculation for individual samples. Can be useful for diagnostics and troubleshooting the model
}
\author{
Vadim Vasilyev
}
