% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_pc_importance}
\alias{plot_pc_importance}
\title{Principal component variance}
\usage{
plot_pc_importance(object)
}
\arguments{
\item{object}{list containing TimeTeller training and test models following \code{train_model}}
}
\value{
Returns the \code{ggplot} object with the summary of cumulative explained variance
}
\description{
Summary of principal component variances explained for the training model
}
\author{
Vadim Vasilyev
}
