% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plot_var_importance}
\alias{plot_var_importance}
\title{Principal component loadings}
\usage{
plot_var_importance(object)
}
\arguments{
\item{object}{list containing TimeTeller training and test models following \code{train_model}}
}
\value{
Returns the \code{ggplot} object with the summary of absolute principal component loadings
}
\description{
Summary of principal component loadings (absolute values) for the training model
}
\author{
Vadim Vasilyev
}
