% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{plotPCs_test}
\alias{plotPCs_test}
\title{Assess precision of time predictions}
\usage{
plotPCs_test(object, ymin = 0, ymax = 24)
}
\arguments{
\item{object}{list containing TimeTeller training and test models following \code{train_model} and \code{test_model} respectively}

\item{ymin}{lower limit for y axis. Helps with visualisation when samples cluster in a particular part of the day (eg biopsy samples taken mostly during day time)}

\item{ymax}{upper limit for y axis}
}
\value{
Returns test data projection onto the first 4 PCs vs estimated time
}
\description{
Useful when assessing the structure of the test data and the precision of \code{TimeTeller} phase estimates
}
\author{
Vadim Vasilyev
}
