% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing_functions.R
\name{shift_outlier_times}
\alias{shift_outlier_times}
\title{Shift predicted phases}
\usage{
shift_outlier_times(object, minT = 8, maxT = 20)
}
\arguments{
\item{object}{list containing TimeTeller training and testing models}

\item{minT}{lower bound for defined time window}

\item{maxT}{upper bound for defined time window}
}
\value{
Returned is the updated object with corrected timing information
}
\description{
Shifts TimeTeller predicted times to more realistic time window IF there is a second likelihood peak located in that window
}
\author{
Vadim Vasilyev
}
