% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{test_model}
\alias{test_model}
\title{Project test data on the training model}
\usage{
test_model(
  object,
  exp_matrix,
  test_grouping_vars,
  test_group_1,
  test_group_2,
  test_group_3,
  test_replicate,
  test_time,
  mat_normalised_test = TRUE,
  log_thresh,
  parallel_comp = FALSE,
  cores = 4,
  minpeakheight = -Inf,
  minpeakdistance = 1,
  nups = 1,
  ndowns = 0,
  threshold = 0,
  npeaks = 2,
  diagnose_pd = FALSE
)
}
\arguments{
\item{object}{object of class list containing timeteller training model (obtained after running \code{train_model} function)}

\item{exp_matrix}{matrix or data frame containing test data with features in rows and samples in columns}

\item{test_grouping_vars}{groups below that will be used for \code{timecourse_matched} normalisation. See the \code{vignette} for examples}

\item{test_group_1}{vector containing metadata for each sample (eg individual, organ). These are not required for \code{intergene} normalisation, however may be useful for visualisation. For \code{timecourse} and \code{timecourse_matched} normalisations, however, these are important and should be supplied in the same format as for the training data to ensure consistent results}

\item{test_group_2}{vector containing metadata for each sample (eg individual, organ)}

\item{test_group_3}{vector containing metadata for each sample (eg individual, organ)}

\item{test_replicate}{vector containing replicate information if available}

\item{test_time}{vector containing replicate information if available. This does not affect the results but is useful to calculate the prediction error and hence the algorithm accuracy}

\item{mat_normalised_test}{was the expression matrix supplied already normalised (eg RMA for microarray, CPM/TPM for RNA, etc) or not. If not, \code{edgeR::cpm} function will be used (pseudocount of 2) to obtain log normalised values for downstream analysis. Default is TRUE}

\item{log_thresh}{log threshold selected. This is an important parameter and should be chosen carefully.Please read \code{help(choose_logthresh_plot)} for further information}

\item{parallel_comp}{if TRUE, parallel computation using \code{foreach} and \code{doParallel} packages will be used. Default is FALSE}

\item{cores}{if using parallel computation, how many cores should be used. Default is 4}

\item{minpeakheight}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development. Please consult \code{?pracma::findpeaks}, if needed, for further information}

\item{minpeakdistance}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{nups}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{ndowns}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{threshold}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{npeaks}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{diagnose_pd}{if TRUE, prints the proportion of interpolated covariance matrices that required nearPD correction. Default is FALSE}
}
\value{
Returned is the rich object of class \code{list} containing the results for both train and test models.
}
\description{
Projects the test data onto the model obtained after running \code{train_model} function.
Among the outputs are the estimated time for each sample and the clock dysfunction metric (Theta)
}
\author{
Vadim Vasilyev
}
