% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeteller.R
\docType{package}
\name{TimeTeller}
\alias{TimeTeller-package}
\alias{TimeTeller}
\title{TimeTeller: Inferring circadian phase and dysfunction from single samples}
\description{
Set of tools and visualisations for interrogating the circadian clock
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VadimVasilyev1994/TimeTeller-v2}
  \item Report bugs at \url{https://github.com/VadimVasilyev1994/TimeTeller-v2/issues}
}

}
\author{
\strong{Maintainer}: Vadim Vasilyev \email{vadim.vasilyev@warwick.ac.uk}

}
