% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{train_model}
\alias{train_model}
\title{Train the TimeTeller model}
\usage{
train_model(
  exp_matrix,
  genes,
  group_1,
  group_2,
  group_3,
  time,
  replicate,
  mat_normalised = TRUE,
  treat_independently = TRUE,
  combine_for_norm = FALSE,
  parallel_comp = FALSE,
  cores = 4,
  method = "intergene",
  grouping_vars = c("Group"),
  num_PC = 3,
  log_thresh,
  epsilon = 0.4,
  eta = 0.35,
  cov_estimate = "normal",
  alpha_par = 0.75,
  num_interp_points = 144,
  interp_method = "perpchip",
  cov_path = "spline",
  minpeakheight = -Inf,
  minpeakdistance = 1,
  nups = 1,
  ndowns = 0,
  threshold = 0,
  npeaks = 2,
  diagnose_pd = FALSE
)
}
\arguments{
\item{exp_matrix}{Matrix or data frame with features in rows and samples in columns}

\item{genes}{genes used to train the TimeTeller model}

\item{group_1}{vector containing metadata for each sample (eg individual, organ)}

\item{group_2}{vector containing metadata for each sample (eg individual, organ). Leave empty if unavailable}

\item{group_3}{vector containing metadata for each sample (eg individual, organ). Leave empty if unavailable}

\item{time}{vector containing timing of each sample. Please appreciate the difference between CT and ZT}

\item{replicate}{vector containing replicate information if available}

\item{mat_normalised}{was the expression matrix supplied already normalised (eg RMA for microarray, CPM/TPM for RNA, etc) or not. If not, \code{edgeR::cpm} function will be used (pseudocount of 2) to obtain log normalised values for downstream analysis. Default is TRUE}

\item{treat_independently}{should replicates be averaged (might be useful for technical replicates, not recommended for biological replicates). Default is TRUE}

\item{combine_for_norm}{should replicates be combined for normalisation. If TRUE, replicates will be combined for normalisation purposes (eg in \code{timecourse} normalisation, replicates will be combined and treated as one timeseries). Default is FALSE}

\item{parallel_comp}{if TRUE, parallel computation using \code{foreach} and \code{doParallel} packages will be used. Default is FALSE}

\item{cores}{if using parallel computation, how many cores should be used. Default is 4}

\item{method}{method used for normalisation. Must be one of the \code{'intergene'}, \code{'timecourse'}, \code{'timecourse_matched'}, \code{'clr'}}

\item{grouping_vars}{group variables (eg \code{'group_1'}, \code{'group_2'} or their combination \code{c('group_1','group_2')}) used for \code{timecourse_matched} normalisation}

\item{num_PC}{how many principal components should be used for the analysis. Default is 3}

\item{log_thresh}{log threshold selected. This is an important parameter and should be chosen carefully. Please read \code{help(choose_logthresh_plot)} for further information}

\item{epsilon}{hyperparameter value used in theta calculation. No reason to change this, more information at \url{https://www.biorxiv.org/content/10.1101/2023.03.14.532177v1}. Default is 0.40}

\item{eta}{hyperparameter value used in theta calculation. No reason to change this, more information at \url{https://www.biorxiv.org/content/10.1101/2023.03.14.532177v1}. Default is 0.35}

\item{cov_estimate}{type of covariance estimate used. If \code{'normal'} is selected, \code{Rfast::mvnorm.mle} will be used. If \code{'robust'} is selected, \code{rrcov::CovMcd} will be used instead. This can be useful when there is a lot of training data, some of which is suspected to have high technical noise. Since we don't want to ignore biological noise (eg chronotype) and it's often hard to tell the two apart, default is \code{'normal'}}

\item{alpha_par}{this is following \code{rrcov} package (in case \code{'robust'} was chosen above) and is the parameter controlling the size of the subsets over which the determinant will be minimized. Allowed values are between 0.5 and 1}

\item{num_interp_points}{number of interpolation points to use for the mean spline and covariance matrices. Default is 144 and this corresponds to 24 * 60 / 144 = 10 minutes}

\item{interp_method}{method used for spline interpolation. Must be one of the \code{'perpchip'} (which uses an adapted periodic version of \code{pchip} from \code{pracma} package) or \code{'standard'} (which uses periodic version of standard \code{spline} function from \code{splines} package).}

\item{cov_path}{method for covariance matrix interpolation. Must be one of the \code{'spline'} (in which case covariance matrix will be interpolated element-wise using \code{interp_method} selection) or \code{'fisherrao'} (in which case the geodesics based on the Fisher–Rao metric for Gaussian distributions will be used).}

\item{minpeakheight}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development. Please consult \code{?pracma::findpeaks}, if needed, for further information}

\item{minpeakdistance}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{nups}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{ndowns}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{threshold}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{npeaks}{this is used internally with \code{pracma} package to locate peaks. No reason to change, however included for future testing / development}

\item{diagnose_pd}{if TRUE, prints the proportion of interpolated covariance matrices that required nearPD correction. Useful for assessing interpolation quality. Default is FALSE}
}
\value{
Returned is the rich object of class \code{list} containing the TimeTeller model for further analysis
}
\description{
Trains the model using provided expression matrix and metadata.
}
\examples{
library("TimeTeller")
tt_model <- train_model(exp_matrix = bjarn_data$expr_mat, genes = bjarn_data$probes_used,
                        group_1 = bjarn_data$group, time = bjarn_data$time, log_thresh = -5)
}
\references{
Vlachou, D., Veretennikova, M., Usselmann, L., Vasilyev, V., Ott, S., Bjarnason, G.A., Dallmann, R., Levi, F. and Rand, D.A., 2023. TimeTeller: a tool to probe the circadian clock as a multigene dynamical system. bioRxiv, pp.2023-03.

Amari, S.I. and Nagaoka, H., 2000. Methods of information geometry (Vol. 191). American Mathematical Soc..
}
\author{
Vadim Vasilyev
}
