% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{choose_genes_tt}
\alias{choose_genes_tt}
\title{Rhythmicity analysis of training data}
\usage{
choose_genes_tt(
  object,
  grouping_var = "Group_1",
  method = "population",
  parallel = TRUE,
  cores = 4
)
}
\arguments{
\item{object}{list containing circadianTT training and test models following \code{train_model} and \code{test_model} respectively}

\item{grouping_var}{how samples should be grouped for rhythmicity analysis. Should be either provided manually or be one of the following: \code{'Group_1'}, \code{'Group_2'}, \code{'Group_3'} or \code{'Replicate'}}

\item{method}{method used for rhythmicity analysis. Default is \code{'population'} as in \url{https://tbiomed.biomedcentral.com/articles/10.1186/1742-4682-11-16}}

\item{parallel}{if TRUE, parallel computation using \code{foreach} and \code{doParallel} packages will be used. Default is TRUE}

\item{cores}{if using parallel computation, how many cores should be used. Default is 4}
}
\value{
Returns plots of theta calculation curves produced using standard \code{graphics::plot}
}
\description{
Calculates rhythmicity results for the training data
}
\references{
Cornelissen, G., 2014. Cosinor-based rhythmometry. Theoretical Biology and Medical Modelling, 11(1), pp.1-24.
}
\author{
Vadim Vasilyev
}
