% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_and_Diagnostics.R
\name{exprs_vs_PredTime_plot}
\alias{exprs_vs_PredTime_plot}
\title{Visualise the the estimated temporal expression of test data}
\usage{
exprs_vs_PredTime_plot(object, genes, theta_thresh, xlim_l = 10, xlim_u = 20)
}
\arguments{
\item{object}{list containing TimeTeller training and test models following \code{train_model} and \code{test_model} respectively}

\item{genes}{genes / features of interest used for plotting}

\item{theta_thresh}{threshold used for theta classification into 'Good' and 'Bad' clocks. This can be subjective and is used for visualisation and sanity check}

\item{xlim_l}{lower limit for x axis. Helps with visualisation when samples cluster in a particular part of the day (eg biopsy samples taken mostly during day time)}

\item{xlim_u}{upper limit for x axis}
}
\value{
Returned is the \code{ggplot} object
}
\description{
Plots both known temporal expression of the training data and estimated temporal expression of the test data. This means that times used for the test data are the ones predicted by \code{TimeTeller}
}
\author{
Vadim Vasilyev
}
