% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeteller.R
\docType{package}
\name{timeteller}
\alias{circadianTT}
\alias{circadianTT-package}
\title{timeteller: Inferring circadian phase and dysfunction from single samples}
\description{
Set of tools and visualisations for interrogating the circadian clock
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VadimVasilyev1994/circadianTT}
  \item Report bugs at \url{https://github.com/VadimVasilyev1994/circadianTT/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
